クラスの整理中。 スキル関係いじっとかないとダメそう？

+++ GameTemplate/Game/monster/parameter/GIManager.cpp
@@ -137,11 +137,11 @@ void GIManager::ActionDicide(int * skill, int * target)
 		return;
 	}
 	int res = rand() % 100;	//適当な乱数。
-	int sum = 0;
+	float sum = 0;
 
 	//行動テーブルをもとに行動させる。
 	for (int i = 0; i < m_AI.size(); i++) {
-		sum += (int)(m_AI[i].rate * 100);
+		sum += m_AI[i].rate * 100;
 		if (sum > res) {
 			*skill = m_AI[i].skillNo;
 			*target = m_AI[i].target;
